Split workflow system tool scope helpers
